// src/types/enums.ts
export enum Location {
  HONG_KONG = "Hong Kong",
  UNITED_STATES = "United States",
  AUSTRALIA = "Australia",
}

export enum Currency {
  HONG_KONG = "HKD",
  UNITED_STATES = "USD",
  AUSTRALIA = "AUD",
}

export enum Package {
  STANDARD = "Standard",
  SAFE = "Safe",
  SUPER_SAFE = "Super Safe",
}